별점 소수점,코멘트500+

+++ ssodal/src/main/java/com/ssodal/dao/BoardReplyDAO.java
@@ -12,7 +12,7 @@ public interface BoardReplyDAO {
 
 	int totalCount(BoardReplyBean brb);
 
-	int avrstars(BoardReplyBean brb);
+	double avrstars(BoardReplyBean brb);
